1.1.0: Use magisk shell instead [1/4]

+++ app/src/main/java/com/xayah/imghelper/Utils/FileUtil.kt
@@ -2,24 +2,16 @@ package com.xayah.imghelper.Utils
 
 import android.content.Context
 import android.util.Log
+import com.topjohnwu.superuser.Shell
 import java.io.File
 import java.io.FileOutputStream
 import java.io.IOException
 
 
-class FileUtil(var context: Context) {
+class FileUtil(private val context: Context) {
 
     fun isFileExist(path: String): Boolean {
-        val mReturn = CommandUtil.executeCommand(
-            "ls $path",
-            "/",
-            true,
-            true
-        )
-        Log.d("FileUtil", "checkFileExist:$mReturn")
-        if (mReturn.contains("No such file or directory")) {
-            return false
-        } else return mReturn.contains(path)
+        return Shell.sh("ls $path").exec().out.isNotEmpty()
     }
 
     fun ifToolsExist(fileName: String): Boolean {
@@ -27,20 +19,27 @@ class FileUtil(var context: Context) {
         return file.exists()
     }
 
-    fun ifFileExist(fileName: String,dirName: String): Boolean {
+    fun ifFileExist(fileName: String, dirName: String): Boolean {
         val file = File(context.filesDir.toString() + "/$dirName/" + fileName)
         return file.exists()
     }
 
-    fun mkDir(path: String) {
-        val file = File(path)
-        if (!file.exists()) {
-            CommandUtil.executeCommand(
-                "mkdir $path",
-                "/system/bin/",
-                true,
-                true
-            )
+
+    fun chmod(path: String, isRoot: Boolean) {
+        if (isRoot) {
+            Shell.su("chmod -R 777 $path").exec()
+        } else {
+            Shell.sh("chmod -R 777 $path").exec()
+        }
+    }
+
+    fun mkDir(path: String, isRoot: Boolean) {
+        if (isRoot) {
+            Shell.su("mkdir $path").exec()
+            chmod(path, true)
+        } else {
+            Shell.sh("mkdir $path").exec()
+            chmod(path, false)
         }
     }
 
@@ -69,13 +68,14 @@ class FileUtil(var context: Context) {
 
     }
 
-    fun cpFiles(pathBefore: String, pathAfter: String) {
-        CommandUtil.executeCommand(
-            "cp $pathBefore $pathAfter",
-            "/",
-            true,
-            true
-        )
+    fun cpFiles(pathBefore: String, pathAfter: String,isRoot: Boolean) {
+        if (isRoot) {
+            Shell.su("cp $pathBefore $pathAfter").exec()
+            chmod(pathAfter, true)
+        } else {
+            Shell.sh("cp $pathBefore $pathAfter").exec()
+            chmod(pathAfter, false)
+        }
     }
 
     fun checkToolsDirExist(dirName: String) {
@@ -85,46 +85,31 @@ class FileUtil(var context: Context) {
         }
     }
 
-    fun moveToolsToFileDir(fileName: String) {
+    fun moveAssetsFileToDir(assetsFilePath: String, fileName: String, dir: String) {
         try {
-            val file = File(context.filesDir.toString() + "/tools/", fileName)
-            val outStream = FileOutputStream(file)
-            val inputStream = context.resources.assets.open("AARCH64/$fileName")
-            val buffer = ByteArray(1024)
-            var byteCount: Int
-            while (inputStream.read(buffer).also { byteCount = it } != -1) { // 循环从输入流读取
-                // buffer字节
-                outStream.write(buffer, 0, byteCount) // 将读取的输入流写入到输出流
+            val file = File(dir, fileName)
+            if (!file.exists()) {
+                val outStream = FileOutputStream(file)
+                val inputStream = context.resources.assets.open(assetsFilePath)
+                val buffer = ByteArray(1024)
+                var byteCount: Int
+                while (inputStream.read(buffer).also { byteCount = it } != -1) { // 循环从输入流读取
+                    // buffer字节
+                    outStream.write(buffer, 0, byteCount) // 将读取的输入流写入到输出流
+                }
+                outStream.flush() // 刷新缓冲区
+                inputStream.close()
+                outStream.close()
+                chmod("$dir/$fileName", false)
             }
-            outStream.flush() // 刷新缓冲区
-            inputStream.close()
-            outStream.close()
         } catch (e: IOException) {
             e.printStackTrace()
         }
     }
 
-    fun moveFileToEnvDir(fileName: String) {
-        try {
-            val file = File(context.filesDir.toString() + "/env/", fileName)
-            val outStream = FileOutputStream(file)
-            val inputStream = context.resources.assets.open("Magisk/$fileName")
-            val buffer = ByteArray(1024)
-            var byteCount: Int
-            while (inputStream.read(buffer).also { byteCount = it } != -1) { // 循环从输入流读取
-                // buffer字节
-                outStream.write(buffer, 0, byteCount) // 将读取的输入流写入到输出流
-            }
-            outStream.flush() // 刷新缓冲区
-            inputStream.close()
-            outStream.close()
-            Log.d("TAG", "moveFileToEnvDir: ")
-        } catch (e: IOException) {
-            e.printStackTrace()
-        }
-    }
 
-    fun moveFileToCertainDir(fileName: String,dirName:String,assetsDirName:String) {
+
+    fun moveFileToCertainDir(fileName: String, dirName: String, assetsDirName: String) {
         try {
             checkToolsDirExist(dirName)
             val file = File(context.filesDir.toString() + "/$dirName/", fileName)
